+++ src/app/api/auth/check/route.js
@@ -1,12 +1,12 @@
+
 import { NextResponse } from "next/server";
 import jwt from 'jsonwebtoken'
 import { cookies } from 'next/headers'
 
-export function GET(request) {
+export function GET() {
   try {
     const cookieStore = cookies()
     const myToken = cookieStore.get('myToken')
-    //const myToken = request.cookies.get('myToken')
     const data = jwt.verify(myToken.value, 'secret')
     return NextResponse.json({id: data.id, nombre: data.nombre, email: data.email,
        dni: data.dni, direccion: data.direccion, telefono: data.telefono, rol: data.rol},
