+++ src/app/api/auth/login/route.js
@@ -35,22 +35,22 @@ export async function POST(request) {
         },
         "secret"
       );
-      const serialized = serialize("myToken", token, {
+      const response = NextResponse.json(
+        {
+          message: "login succesful",
+          rol: usuario.rol,
+        },
+        { status: 200 }
+      )
+      response.cookies.set({
+        name: "myToken",
+        value: token,
+        path: "/",
         httpOnly: true,
         secure: process.env.NODE_ENV === "production",
-        sameSite: "strict",
-        maxAge: 1000 * 3600 * 24 * 30,
-        path: "/",
-      });
-      return new Response(
-        JSON.stringify({ message: "success", rol: usuario.rol }),
-        {
-          status: 200,
-          headers: {
-            "Set-Cookie": serialized,
-          },
-        }
-      );
+        sameSite: "strict"
+      })
+      return response
     } else {
       // Contraseña incorrecta, deniega el acceso
       return NextResponse.json(
@@ -63,5 +63,6 @@ export async function POST(request) {
       message: "Ocurrio un error en el inicio de sesion",
       error: error.message,
     });
+    
   }
 }
